+++ ReportSimulationOutput/tests/output_report_test.rb
@@ -310,137 +310,6 @@ class ReportSimulationOutputTest < MiniTest::Test
     'Surface Construction Index: Window4'
   ]
 
-  ERIRows = [
-    'hpxml_heat_sys_ids',
-    'hpxml_cool_sys_ids',
-    'hpxml_dhw_sys_ids',
-    'hpxml_vent_preheat_sys_ids',
-    'hpxml_vent_precool_sys_ids',
-    'hpxml_eec_heats',
-    'hpxml_eec_cools',
-    'hpxml_eec_dhws',
-    'hpxml_eec_vent_preheats',
-    'hpxml_eec_vent_precools',
-    'hpxml_heat_fuels',
-    'hpxml_dwh_fuels',
-    'hpxml_vent_preheat_fuels',
-    'fuelElectricity',
-    'fuelNaturalGas',
-    'fuelFuelOil',
-    'fuelPropane',
-    'fuelWoodCord',
-    'fuelWoodPellets',
-    'fuelCoal',
-    'enduseElectricityHeating',
-    'enduseElectricityHeatingFansPumps',
-    'enduseElectricityCooling',
-    'enduseElectricityCoolingFansPumps',
-    'enduseElectricityHotWater',
-    'enduseElectricityHotWaterRecircPump',
-    'enduseElectricityHotWaterSolarThermalPump',
-    'enduseElectricityLightingInterior',
-    'enduseElectricityLightingGarage',
-    'enduseElectricityLightingExterior',
-    'enduseElectricityMechVent',
-    'enduseElectricityMechVentPreheating',
-    'enduseElectricityMechVentPrecooling',
-    'enduseElectricityWholeHouseFan',
-    'enduseElectricityRefrigerator',
-    'enduseElectricityDehumidifier',
-    'enduseElectricityDishwasher',
-    'enduseElectricityClothesWasher',
-    'enduseElectricityClothesDryer',
-    'enduseElectricityRangeOven',
-    'enduseElectricityCeilingFan',
-    'enduseElectricityTelevision',
-    'enduseElectricityPlugLoads',
-    'enduseElectricityPV',
-    'enduseElectricityGenerator',
-    'enduseNaturalGasHeating',
-    'enduseNaturalGasHotWater',
-    'enduseNaturalGasClothesDryer',
-    'enduseNaturalGasRangeOven',
-    'enduseNaturalGasMechVentPreheating',
-    'enduseNaturalGasGenerator',
-    'enduseFuelOilHeating',
-    'enduseFuelOilHotWater',
-    'enduseFuelOilClothesDryer',
-    'enduseFuelOilRangeOven',
-    'enduseFuelOilMechVentPreheating',
-    'enduseFuelOilGenerator',
-    'endusePropaneHeating',
-    'endusePropaneHotWater',
-    'endusePropaneClothesDryer',
-    'endusePropaneRangeOven',
-    'endusePropaneMechVentPreheating',
-    'endusePropaneGenerator',
-    'enduseWoodCordHeating',
-    'enduseWoodCordHotWater',
-    'enduseWoodCordClothesDryer',
-    'enduseWoodCordRangeOven',
-    'enduseWoodCordMechVentPreheating',
-    'enduseWoodCordGenerator',
-    'enduseWoodPelletsHeating',
-    'enduseWoodPelletsHotWater',
-    'enduseWoodPelletsClothesDryer',
-    'enduseWoodPelletsRangeOven',
-    'enduseWoodPelletsMechVentPreheating',
-    'enduseWoodPelletsGenerator',
-    'enduseCoalHeating',
-    'enduseCoalHotWater',
-    'enduseCoalClothesDryer',
-    'enduseCoalRangeOven',
-    'enduseCoalMechVentPreheating',
-    'enduseCoalGenerator',
-    'loadHeatingDelivered',
-    'loadCoolingDelivered',
-    'loadHotWaterDelivered',
-    'co2eCambiumHourlyMidCaseLRMERRMPATotal',
-    'co2eCambiumHourlyMidCaseLRMERRMPAElectricity',
-    'co2eCambiumHourlyMidCaseLRMERRMPANaturalGas',
-    'co2eCambiumHourlyMidCaseLRMERRMPAFuelOil',
-    'co2eCambiumHourlyMidCaseLRMERRMPAPropane',
-    'co2eCambiumHourlyMidCaseLRMERRMPAWoodCord',
-    'co2eCambiumHourlyMidCaseLRMERRMPAWoodPellets',
-    'co2eCambiumHourlyMidCaseLRMERRMPACoal',
-    'co2eCambiumHourlyLowRECostsLRMERRMPATotal',
-    'co2eCambiumHourlyLowRECostsLRMERRMPAElectricity',
-    'co2eCambiumHourlyLowRECostsLRMERRMPANaturalGas',
-    'co2eCambiumHourlyLowRECostsLRMERRMPAFuelOil',
-    'co2eCambiumHourlyLowRECostsLRMERRMPAPropane',
-    'co2eCambiumHourlyLowRECostsLRMERRMPAWoodCord',
-    'co2eCambiumHourlyLowRECostsLRMERRMPAWoodPellets',
-    'co2eCambiumHourlyLowRECostsLRMERRMPACoal',
-    'co2eCambiumAnnualMidCaseAERNationalTotal',
-    'co2eCambiumAnnualMidCaseAERNationalElectricity',
-    'co2eCambiumAnnualMidCaseAERNationalNaturalGas',
-    'co2eCambiumAnnualMidCaseAERNationalFuelOil',
-    'co2eCambiumAnnualMidCaseAERNationalPropane',
-    'co2eCambiumAnnualMidCaseAERNationalWoodCord',
-    'co2eCambiumAnnualMidCaseAERNationalWoodPellets',
-    'co2eCambiumAnnualMidCaseAERNationalCoal',
-    'so2eGRIDRMPATotal',
-    'so2eGRIDRMPAElectricity',
-    'so2eGRIDRMPANaturalGas',
-    'so2eGRIDRMPAFuelOil',
-    'so2eGRIDRMPAPropane',
-    'so2eGRIDRMPAWoodCord',
-    'so2eGRIDRMPAWoodPellets',
-    'so2eGRIDRMPACoal',
-    'noxeGRIDRMPATotal',
-    'noxeGRIDRMPAElectricity',
-    'noxeGRIDRMPANaturalGas',
-    'noxeGRIDRMPAFuelOil',
-    'noxeGRIDRMPAPropane',
-    'noxeGRIDRMPAWoodCord',
-    'noxeGRIDRMPAWoodPellets',
-    'noxeGRIDRMPACoal',
-    'hpxml_cfa',
-    'hpxml_nbr',
-    'hpxml_nst',
-    'hpxml_residential_facility_type',
-  ]
-
   def all_base_hpxml_timeseries_cols
     return (BaseHPXMLTimeseriesColsEnergy +
             BaseHPXMLTimeseriesColsFuels +
@@ -453,22 +322,76 @@ class ReportSimulationOutputTest < MiniTest::Test
             BaseHPXMLTimeseriesColsWeather)
   end
 
+  def emission_scenarios
+    return ['CO2e: Cambium Hourly MidCase LRMER RMPA',
+            'CO2e: Cambium Hourly LowRECosts LRMER RMPA',
+            'CO2e: Cambium Annual MidCase AER National',
+            'SO2: eGRID RMPA',
+            'NOx: eGRID RMPA']
+  end
+
+  def emission_annual_cols
+    cols = []
+    emission_scenarios.each do |scenario|
+      cols += ["Emissions: #{scenario}: Total (lb)",
+               "Emissions: #{scenario}: Electricity: Total (lb)",
+               "Emissions: #{scenario}: Electricity: Heating Fans/Pumps (lb)",
+               "Emissions: #{scenario}: Electricity: Cooling (lb)",
+               "Emissions: #{scenario}: Electricity: Cooling Fans/Pumps (lb)",
+               "Emissions: #{scenario}: Electricity: Hot Water (lb)",
+               "Emissions: #{scenario}: Electricity: Lighting Interior (lb)",
+               "Emissions: #{scenario}: Electricity: Lighting Exterior (lb)",
+               "Emissions: #{scenario}: Electricity: Refrigerator (lb)",
+               "Emissions: #{scenario}: Electricity: Dishwasher (lb)",
+               "Emissions: #{scenario}: Electricity: Clothes Washer (lb)",
+               "Emissions: #{scenario}: Electricity: Clothes Dryer (lb)",
+               "Emissions: #{scenario}: Electricity: Range/Oven (lb)",
+               "Emissions: #{scenario}: Electricity: Television (lb)",
+               "Emissions: #{scenario}: Electricity: Plug Loads (lb)",
+               "Emissions: #{scenario}: Electricity: PV (lb)",
+               "Emissions: #{scenario}: Natural Gas: Total (lb)",
+               "Emissions: #{scenario}: Natural Gas: Heating (lb)"]
+    end
+    return cols
+  end
+
   def emissions_timeseries_cols
-    return ['Emissions: CO2e: Cambium Hourly MidCase LRMER RMPA: Total',
-            'Emissions: CO2e: Cambium Hourly MidCase LRMER RMPA: Electricity',
-            'Emissions: CO2e: Cambium Hourly MidCase LRMER RMPA: Natural Gas',
-            'Emissions: CO2e: Cambium Hourly LowRECosts LRMER RMPA: Total',
-            'Emissions: CO2e: Cambium Hourly LowRECosts LRMER RMPA: Electricity',
-            'Emissions: CO2e: Cambium Hourly LowRECosts LRMER RMPA: Natural Gas',
-            'Emissions: CO2e: Cambium Annual MidCase AER National: Total',
-            'Emissions: CO2e: Cambium Annual MidCase AER National: Electricity',
-            'Emissions: CO2e: Cambium Annual MidCase AER National: Natural Gas',
-            'Emissions: SO2: eGRID RMPA: Total',
-            'Emissions: SO2: eGRID RMPA: Electricity',
-            'Emissions: SO2: eGRID RMPA: Natural Gas',
-            'Emissions: NOx: eGRID RMPA: Total',
-            'Emissions: NOx: eGRID RMPA: Electricity',
-            'Emissions: NOx: eGRID RMPA: Natural Gas']
+    cols = []
+    emission_scenarios.each do |scenario|
+      cols += ["Emissions: #{scenario}: Total"]
+    end
+    return cols
+  end
+
+  def emission_fuels_timeseries_cols
+    cols = []
+    emission_scenarios.each do |scenario|
+      cols += ["Emissions: #{scenario}: Electricity: Total",
+               "Emissions: #{scenario}: Natural Gas: Total"]
+    end
+    return cols
+  end
+
+  def emission_end_uses_timeseries_cols
+    cols = []
+    emission_scenarios.each do |scenario|
+      cols += ["Emissions: #{scenario}: Electricity: Heating Fans/Pumps",
+               "Emissions: #{scenario}: Electricity: Cooling",
+               "Emissions: #{scenario}: Electricity: Cooling Fans/Pumps",
+               "Emissions: #{scenario}: Electricity: Hot Water",
+               "Emissions: #{scenario}: Electricity: Lighting Interior",
+               "Emissions: #{scenario}: Electricity: Lighting Exterior",
+               "Emissions: #{scenario}: Electricity: Refrigerator",
+               "Emissions: #{scenario}: Electricity: Dishwasher",
+               "Emissions: #{scenario}: Electricity: Clothes Washer",
+               "Emissions: #{scenario}: Electricity: Clothes Dryer",
+               "Emissions: #{scenario}: Electricity: Range/Oven",
+               "Emissions: #{scenario}: Electricity: Television",
+               "Emissions: #{scenario}: Electricity: Plug Loads",
+               "Emissions: #{scenario}: Electricity: PV",
+               "Emissions: #{scenario}: Natural Gas: Heating"]
+    end
+    return cols
   end
 
   def test_annual_only
@@ -479,6 +402,8 @@ class ReportSimulationOutputTest < MiniTest::Test
                   'include_timeseries_fuel_consumptions' => false,
                   'include_timeseries_end_use_consumptions' => false,
                   'include_timeseries_emissions' => false,
+                  'include_timeseries_emission_fuels' => false,
+                  'include_timeseries_emission_end_uses' => false,
                   'include_timeseries_hot_water_uses' => false,
                   'include_timeseries_total_loads' => false,
                   'include_timeseries_unmet_hours' => false,
@@ -495,13 +420,15 @@ class ReportSimulationOutputTest < MiniTest::Test
   end
 
   def test_annual_only2
-    args_hash = { 'hpxml_path' => '../workflow/sample_files/base.xml',
+    args_hash = { 'hpxml_path' => '../workflow/sample_files/base-misc-emissions.xml',
                   'add_component_loads' => true,
                   'timeseries_frequency' => 'none',
                   'include_timeseries_total_consumptions' => false,
                   'include_timeseries_fuel_consumptions' => true,
                   'include_timeseries_end_use_consumptions' => true,
                   'include_timeseries_emissions' => true,
+                  'include_timeseries_emission_fuels' => true,
+                  'include_timeseries_emission_end_uses' => true,
                   'include_timeseries_hot_water_uses' => true,
                   'include_timeseries_total_loads' => true,
                   'include_timeseries_unmet_hours' => true,
@@ -512,7 +439,7 @@ class ReportSimulationOutputTest < MiniTest::Test
     annual_csv, timeseries_csv = _test_measure(args_hash)
     assert(File.exist?(annual_csv))
     assert(!File.exist?(timeseries_csv))
-    expected_annual_rows = AnnualRows
+    expected_annual_rows = AnnualRows + emission_annual_cols
     actual_annual_rows = File.readlines(annual_csv).map { |x| x.split(',')[0].strip }.select { |x| !x.empty? }
     assert_equal(expected_annual_rows.sort, actual_annual_rows.sort)
   end
@@ -602,6 +529,40 @@ class ReportSimulationOutputTest < MiniTest::Test
     timeseries_cols = timeseries_rows.transpose
     _check_for_constant_timeseries_step(timeseries_cols[0])
     _check_for_nonzero_timeseries_value(timeseries_csv, emissions_timeseries_cols[0..2])
+  end
+
+  def test_timeseries_hourly_emission_end_uses
+    args_hash = { 'hpxml_path' => '../workflow/sample_files/base-misc-emissions.xml',
+                  'timeseries_frequency' => 'hourly',
+                  'include_timeseries_emission_end_uses' => true }
+    annual_csv, timeseries_csv = _test_measure(args_hash)
+    assert(File.exist?(annual_csv))
+    assert(File.exist?(timeseries_csv))
+    expected_timeseries_cols = ['Time'] + emission_end_uses_timeseries_cols
+    actual_timeseries_cols = File.readlines(timeseries_csv)[0].strip.split(',')
+    assert_equal(expected_timeseries_cols.sort, actual_timeseries_cols.sort)
+    timeseries_rows = CSV.read(timeseries_csv)
+    assert_equal(8760, timeseries_rows.size - 2)
+    timeseries_cols = timeseries_rows.transpose
+    _check_for_constant_timeseries_step(timeseries_cols[0])
+    _check_for_nonzero_timeseries_value(timeseries_csv, emission_end_uses_timeseries_cols[0..2])
+  end
+
+  def test_timeseries_hourly_emission_fuels
+    args_hash = { 'hpxml_path' => '../workflow/sample_files/base-misc-emissions.xml',
+                  'timeseries_frequency' => 'hourly',
+                  'include_timeseries_emission_fuels' => true }
+    annual_csv, timeseries_csv = _test_measure(args_hash)
+    assert(File.exist?(annual_csv))
+    assert(File.exist?(timeseries_csv))
+    expected_timeseries_cols = ['Time'] + emission_fuels_timeseries_cols
+    actual_timeseries_cols = File.readlines(timeseries_csv)[0].strip.split(',')
+    assert_equal(expected_timeseries_cols.sort, actual_timeseries_cols.sort)
+    timeseries_rows = CSV.read(timeseries_csv)
+    assert_equal(8760, timeseries_rows.size - 2)
+    timeseries_cols = timeseries_rows.transpose
+    _check_for_constant_timeseries_step(timeseries_cols[0])
+    _check_for_nonzero_timeseries_value(timeseries_csv, emission_fuels_timeseries_cols[0..2])
   end
 
   def test_timeseries_hourly_enduses
@@ -860,6 +821,8 @@ class ReportSimulationOutputTest < MiniTest::Test
                   'include_timeseries_fuel_consumptions' => true,
                   'include_timeseries_end_use_consumptions' => true,
                   'include_timeseries_emissions' => true,
+                  'include_timeseries_emission_fuels' => true,
+                  'include_timeseries_emission_end_uses' => true,
                   'include_timeseries_hot_water_uses' => true,
                   'include_timeseries_total_loads' => true,
                   'include_timeseries_unmet_hours' => true,
@@ -870,7 +833,11 @@ class ReportSimulationOutputTest < MiniTest::Test
     annual_csv, timeseries_csv = _test_measure(args_hash)
     assert(File.exist?(annual_csv))
     assert(File.exist?(timeseries_csv))
-    expected_timeseries_cols = ['Time'] + all_base_hpxml_timeseries_cols + emissions_timeseries_cols +
+    expected_timeseries_cols = ['Time'] +
+                               all_base_hpxml_timeseries_cols +
+                               emissions_timeseries_cols +
+                               emission_fuels_timeseries_cols +
+                               emission_end_uses_timeseries_cols +
                                ['End Use: Electricity: PV', 'Energy Use: Net', 'Fuel Use: Electricity: Net']
     actual_timeseries_cols = File.readlines(timeseries_csv)[0].strip.split(',')
     assert_equal(expected_timeseries_cols.sort, actual_timeseries_cols.sort)
@@ -889,6 +856,8 @@ class ReportSimulationOutputTest < MiniTest::Test
                   'include_timeseries_fuel_consumptions' => true,
                   'include_timeseries_end_use_consumptions' => true,
                   'include_timeseries_emissions' => true,
+                  'include_timeseries_emission_fuels' => true,
+                  'include_timeseries_emission_end_uses' => true,
                   'include_timeseries_hot_water_uses' => true,
                   'include_timeseries_total_loads' => true,
                   'include_timeseries_unmet_hours' => true,
@@ -899,7 +868,11 @@ class ReportSimulationOutputTest < MiniTest::Test
     annual_csv, timeseries_csv = _test_measure(args_hash)
     assert(File.exist?(annual_csv))
     assert(File.exist?(timeseries_csv))
-    expected_timeseries_cols = ['Time'] + all_base_hpxml_timeseries_cols + emissions_timeseries_cols +
+    expected_timeseries_cols = ['Time'] +
+                               all_base_hpxml_timeseries_cols +
+                               emissions_timeseries_cols +
+                               emission_fuels_timeseries_cols +
+                               emission_end_uses_timeseries_cols +
                                ['End Use: Electricity: PV', 'Energy Use: Net', 'Fuel Use: Electricity: Net']
     actual_timeseries_cols = File.readlines(timeseries_csv)[0].strip.split(',')
     assert_equal(expected_timeseries_cols.sort, actual_timeseries_cols.sort)
@@ -918,6 +891,8 @@ class ReportSimulationOutputTest < MiniTest::Test
                   'include_timeseries_fuel_consumptions' => true,
                   'include_timeseries_end_use_consumptions' => true,
                   'include_timeseries_emissions' => true,
+                  'include_timeseries_emission_fuels' => true,
+                  'include_timeseries_emission_end_uses' => true,
                   'include_timeseries_hot_water_uses' => true,
                   'include_timeseries_total_loads' => true,
                   'include_timeseries_unmet_hours' => true,
@@ -928,7 +903,11 @@ class ReportSimulationOutputTest < MiniTest::Test
     annual_csv, timeseries_csv = _test_measure(args_hash)
     assert(File.exist?(annual_csv))
     assert(File.exist?(timeseries_csv))
-    expected_timeseries_cols = ['Time'] + all_base_hpxml_timeseries_cols + emissions_timeseries_cols +
+    expected_timeseries_cols = ['Time'] +
+                               all_base_hpxml_timeseries_cols +
+                               emissions_timeseries_cols +
+                               emission_fuels_timeseries_cols +
+                               emission_end_uses_timeseries_cols +
                                ['End Use: Electricity: PV', 'Energy Use: Net', 'Fuel Use: Electricity: Net']
     actual_timeseries_cols = File.readlines(timeseries_csv)[0].strip.split(',')
     assert_equal(expected_timeseries_cols.sort, actual_timeseries_cols.sort)
@@ -957,6 +936,8 @@ class ReportSimulationOutputTest < MiniTest::Test
                   'include_timeseries_fuel_consumptions' => false,
                   'include_timeseries_end_use_consumptions' => false,
                   'include_timeseries_emissions' => true,
+                  'include_timeseries_emission_fuels' => true,
+                  'include_timeseries_emission_end_uses' => true,
                   'include_timeseries_hot_water_uses' => false,
                   'include_timeseries_total_loads' => false,
                   'include_timeseries_unmet_hours' => false,
@@ -967,7 +948,10 @@ class ReportSimulationOutputTest < MiniTest::Test
     annual_csv, timeseries_csv = _test_measure(args_hash)
     assert(File.exist?(annual_csv))
     assert(File.exist?(timeseries_csv))
-    expected_timeseries_cols = ['Time'] + emissions_timeseries_cols
+    expected_timeseries_cols = ['Time'] +
+                               emissions_timeseries_cols +
+                               emission_fuels_timeseries_cols +
+                               emission_end_uses_timeseries_cols
     actual_timeseries_cols = File.readlines(timeseries_csv)[0].strip.split(',')
     assert_equal(expected_timeseries_cols.sort, actual_timeseries_cols.sort)
     timeseries_rows = CSV.read(timeseries_csv)
@@ -1125,6 +1109,8 @@ class ReportSimulationOutputTest < MiniTest::Test
                   'include_timeseries_fuel_consumptions' => false,
                   'include_timeseries_end_use_consumptions' => false,
                   'include_timeseries_emissions' => false,
+                  'include_timeseries_emission_fuels' => false,
+                  'include_timeseries_emission_end_uses' => false,
                   'include_timeseries_hot_water_uses' => false,
                   'include_timeseries_total_loads' => false,
                   'include_timeseries_unmet_hours' => false,
@@ -1154,6 +1140,8 @@ class ReportSimulationOutputTest < MiniTest::Test
                   'include_timeseries_fuel_consumptions' => false,
                   'include_timeseries_end_use_consumptions' => false,
                   'include_timeseries_emissions' => false,
+                  'include_timeseries_emission_fuels' => false,
+                  'include_timeseries_emission_end_uses' => false,
                   'include_timeseries_hot_water_uses' => false,
                   'include_timeseries_total_loads' => false,
                   'include_timeseries_unmet_hours' => false,